% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_fact.R
\name{cat_fact}
\alias{cat_fact}
\title{get a cat fact}
\usage{
cat_fact(max_length = NULL, ...)
}
\arguments{
\item{max_length}{(integer) limits the length of the facts returned}

\item{...}{curl options passed on to \code{httr::GET}}
}
\value{
a cat fact in a list
}
\description{
get a cat fact
}
\examples{
\dontrun{
cat_fact()
cat_fact(40)
cat_fact(80)
cat_fact('asdf') # invalid max_length returns nothing
}
}
